Vulkan: Remove false memory accesses in non-updated coherent memory
This can occur when memory is mapped coherently, but gapii doesn't observe any
application writes to the mapped memory.

Recording a WRITE to the device memory regions that weren't not actually updated
is problematic since it can hide writes by earlier commands that actually did
write to these memory regions.