Graph Visualization: Added a faster function to get the pbtxt file and fixed the super-node(hierarchy) based on frames.
Faster function to get pbtxt: Now it takes only ~45 seconds to get the
pbtxt file from multithreading trace with 50 frames, but it is not
possible visualize the graph in TensorBoard due thousand of nodes
displayed in the top level.

Hieararchy by Frame: It was implemented running a depth first search for
every QueuePresent command (node). Now it is possible visualize the top
level of the graph, but not yet lower levels.