Graph visualization: Added General Chunk Algorithm , restructuration of the code into four components and new representation of Label.
General Chunk Algorithm: This algorithm creates super-nodes to avoid
display more than a fix number of nodes in every level. For every
hierarchy, the nodes are grouped in super-nodes based on command names,
then if the numbers of nodes exceed the maximim limit, chunks are
created, and if in the new chunks the numbers of nodes exceed the maximmum limit,
then new chunks are created and so on.

Restructuration of the code: The code is divided now in four
components: 1) graph_structure: Create and manage the data structure for
the graph efficiently. 2) graph_algorihtms: Contains the algorihtms such
as: General Chunk Algorithm, deph first search , etc. 3) graph_output:
Code to obtain the graph in differents format: pbtxt, proto and dot and
4) graph_visualization: Receive the trace , process the dependency graph
into own data structure and return the output file in the desired format.

New representation of Label: To apply efficiently the chunk algorithm
and to adreess complexity problems in time and memory, label now is
defined like two slices of names and ids.

Results: Using the General Chunk Algorithm is possible to visualize any
trace such as: Macaroon with 3 frames (~7 minutes to obtain the pbtxt
file, ~15 seconds to upload in TensorBoard and ~2secons to expand
super-nodes).

Note1: All "op" attribute for nodes in pbtxt must be different, since the find
similar subgraphs algorithm from TensorBoard compute the similarity
between  nodes with same "op".

Note2: Adding colors to the nodes increase the time for expand super-nodes.